Print notification to file functionality (for scripting).
Also organized some other stuff.
