+++ src/cli.rs
@@ -1,14 +1,129 @@
 use std::io::Write;
-use std::os::unix::net::UnixStream;
+use std::os::unix::net::{UnixStream, UnixListener};
 use std::panic::panic_any;
+use std::io::{BufRead, BufReader};
+use std::path::Path;
 
 use getopts::Options;
+use winit::window::WindowId;
+use winit::event_loop::EventLoopWindowTarget;
+
+use crate::NotifyWindowManager;
+
+pub const SOCKET_PATH: &str = "/tmp/wired.sock";
+
+#[derive(Debug)]
+pub enum SocketError {
+    Parse(&'static str),
+    NotificationNotFound,
+    InvalidCommand,
+}
 
 pub enum ShouldRun {
     Yes,
     No,
 }
 
+pub fn init_socket_listener() -> Result<UnixListener, String> {
+    // Socket, for listening to CLI calls to ourselves.
+    // We leave the socket up in pretty much all cases when closing, and just unbind it always.
+    // This could cause confusing behavior for users where they have 2 wired instances running, and
+    // neither will work properly (one will have the notification bus name, and one will have the
+    // socket).
+    // "Fixing" this would require making sure the socket is unbound in almost all cases -- likely
+    // having to import a few crates like Ctrl-C and others -- yuck.
+    // Let's just leave it as it is and try to communicate to users that it's not an issue.
+    // https://stackoverflow.com/questions/40218416/how-do-i-close-a-unix-socket-in-rust
+    let socket_path = Path::new(SOCKET_PATH);
+    if socket_path.exists() {
+        println!(
+            "A wired socket exists; taking ownership.  Existing wired processes will not receive CLI calls."
+        );
+        std::fs::remove_file(SOCKET_PATH).unwrap();
+    }
+    let listener = match UnixListener::bind(socket_path) {
+        Ok(sock) => sock,
+        Err(e) => {
+            return Err(format!("Couldn't bind socket {}\n{:?}", SOCKET_PATH, e));
+        }
+    };
+    listener.set_nonblocking(true).unwrap();
+    Ok(listener)
+}
+
+// Socket stuff:
+fn get_window_id(arg: &str, manager: &NotifyWindowManager) -> Result<WindowId, SocketError> {
+    let idx = if arg == "latest" {
+        let num_windows = manager.monitor_windows.values().flatten().count();
+        if num_windows > 0 {
+            num_windows - 1
+        } else {
+            return Err(SocketError::NotificationNotFound)
+        }
+    } else if let Ok(idx) = arg.parse::<usize>() {
+        idx
+    } else {
+        return Err(SocketError::Parse("Value is not of type usize."));
+    };
+
+    if let Some(window) = manager.find_window_ordered(idx) {
+        Ok(window)
+    } else {
+        Err(SocketError::NotificationNotFound)
+    }
+}
+
+pub fn handle_socket_message(manager: &mut NotifyWindowManager, el: &EventLoopWindowTarget<()>, stream: UnixStream) -> Result<(), SocketError> {
+    let stream = BufReader::new(stream);
+    for line in stream.lines() {
+        let line = match line {
+            Ok(l) => l,
+            Err(_) => continue,
+        };
+
+        println!("Recived socket message: {}", line);
+        if let Some((command, args)) = line.split_once(":") {
+            match command {
+                "close" => {
+                    if args == "all" {
+                        manager.drop_windows();
+                    } else {
+                        let id = get_window_id(args, manager)?;
+                        manager.drop_window_id(id);
+                    }
+                }
+                "action" => {
+                    let (notif_id, action_id) =
+                        args.split_once(",")
+                            .ok_or(SocketError::Parse("Malformed action request."))?;
+
+                    let id = get_window_id(notif_id, manager)?;
+                    let action = action_id.parse::<usize>()
+                        .map_err(|_| SocketError::Parse("Value is not of type usize."))?;
+                    manager.trigger_action_idx(id, action);
+                },
+                "show" => {
+                    let num = args.parse::<usize>()
+                        .map_err(|_| SocketError::Parse("Value is not of type usize."))?;
+                    for _ in 0..num {
+                        if let Some(n) = manager.history.pop_back() {
+                            manager.new_notification(n, el);
+                        }
+                    }
+                }
+                _ => {
+                    return Err(SocketError::InvalidCommand)
+                },
+            }
+        } else {
+            return Err(SocketError::Parse("Malformed command."));
+        }
+    }
+
+    Ok(())
+}
+
+// CLI stuff:
 fn print_usage(opts: Options) {
     print!("{}", opts.usage("Usage:\twired [options]\n\tIDX refers to the Nth most recent notification."));
 }
@@ -69,7 +184,7 @@ pub fn process_cli(args: Vec<String>) -> Result<ShouldRun, String> {
 
     // All these options use a socket.
     if matches.opt_present("d") || matches.opt_present("a") || matches.opt_present("s") {
-        let mut sock = match UnixStream::connect(crate::SOCKET_PATH) {
+        let mut sock = match UnixStream::connect(SOCKET_PATH) {
             Ok(s) => s,
             Err(e) => {
                 return Err(format!(
